Added Service Tests

+++ backend/src/main/java/com/exercise/api/service/ExerciseService.java
@@ -19,23 +19,38 @@ public class ExerciseService {
         return exercises;
     }
 
-    public Exercise getExerciseById(long id) {
+    public Exercise getExerciseById(long id) throws Exception {
+        if (id < 0) {
+            throw new Exception("exerciseId cannot be less than 0");
+        }
         return exerciseRepository.findById(id).get();
     }
 
-    public void saveOrUpdate(Exercise exercise) {
+    public void saveOrUpdate(Exercise exercise) throws Exception {
+        if (exercise.getExerciseId() < 0) {
+            throw new Exception("exerciseID cannot be less than 0");
+        }
         exerciseRepository.save(exercise);
     }
 
-    public void delete(long id) {
+    public void delete(long id) throws Exception {
+        if (id < 0) {
+            throw new Exception("exerciseID cannot be less than 0");
+        }
         exerciseRepository.deleteById(id);
     }
 
-    public List<Exercise> getExerciseByWorkoutId(long workoutId) {
+    public List<Exercise> getExerciseByWorkoutId(long workoutId) throws Exception {
+        if (workoutId < 0) {
+            throw new Exception("workoutID cannot be less than 0");
+        }
         return exerciseRepository.findExerciseByWorkoutId(workoutId);
     }
 
-    public List<Exercise> getExerciseByWorkoutIdAndType(long workoutId, String type) {
+    public List<Exercise> getExerciseByWorkoutIdAndType(long workoutId, String type) throws Exception{
+        if (workoutId < 0) {
+            throw new Exception("workoutID cannot be less than 0");
+        }
         return exerciseRepository.findExerciseByWorkoutIdAndType(workoutId, type);
     }
 }
